# Use mmap for allocating heap pages in the GC (try 2)