This fixes the Solaris incompatibility in #4221. I have confirmed that this branch passes the bootstrap tests on Solaris.

The diff between this branch and #4221 is as follows:

```diff
diff --git a/gc.c b/gc.c
index 05555adb5a..df9779e01d 100644
--- a/gc.c
+++ b/gc.c
@@ -10392,7 +10392,7 @@ rb_aligned_malloc(size_t alignment, size_t size)
 #elif defined(HAVE_MMAP)
     GC_ASSERT(alignment % sysconf(_SC_PAGE_SIZE) == 0);

-    char *ptr = mmap(NULL, alignment + size, PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANONYMOUS, 0, 0);
+    char *ptr = mmap(NULL, alignment + size, PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
     if (ptr == MAP_FAILED) {
         return NULL;
     }
```

The problem was that according to the documentation for `mmap`:

> The fd argument is ignored; however, some implementations require fd to be -1 if MAP_ANONYMOUS (or MAP_ANON) is specified, and portable applications should ensure this.

# Original PR description

This PR serves two purposes:

1. `mmap` OS pages directly in the GC heap, this gives a performance gain
2. Fix a bug with auto compaction due to a failing `mprotect` call (ticket: https://bugs.ruby-lang.org/issues/17652).

# Performance

Calling `mmap` directly (over `posix_memalign`) gives a small performance gain, as shown in the benchmarks below.

```
+-------------------+--------------------------+--------------------------+---------------+
| Name              | Base score               | Feature score            | % improvement |
+-------------------+--------------------------+--------------------------+---------------+
| liquid parse      | 37.02 i/s (± 0.22%)      | 37.60 i/s (± 0.19%)      | 1.57%         |
| liquid render     | 120.08 i/s (± 0.35%)     | 120.58 i/s (± 0.23%)     | 0.42%         |
| optcarrot fps     | 41.45 fps (± 0.26%)      | 43.24 fps (± 0.14%)      | 4.33%         |
| optcarrot p95 fps | 41.05 fps (± 0.20%)      | 41.80 fps (± 0.03%)      | 1.83%         |
| railsbench        | 101.49 i/s (± 0.45%)     | 101.33 i/s (± 0.29%)     | -0.17%        |
+-------------------+--------------------------+--------------------------+---------------+
```

In the table above, base is Ruby master at 07ab172ebef28bb82c2650694548d123802d6c22 and feature is this branch.

# Auto compaction bug

Redmine ticket: https://bugs.ruby-lang.org/issues/17652

GC compaction will crash on some pages due to a failing `mprotect` call. According to the `mprotect` documentation:

> POSIX says that the behavior of mprotect() is unspecified if it is applied to a region of memory that was not obtained via mmap(2).

This causes `mprotect` to sometimes fail.

## Reproduction

```ruby
GC.auto_compact = true

times = 20_000_000
arr = Array.new(times)
times.times do |i|
  arr[i] = "#{i}"
end

arr = Array.new(1_000_000, 42)
GC.start

puts "ok"
```

## Expected output

Program runs and prints "ok".

## Actual output

Crashes on Ubuntu 18.04. See `crash.log` for the crash log (note that the log file has been truncated because it's too big).